Collate the flash port code for stm32l4xx_cube, and fix the cross-bank issue
